Refactor post-auth0 support.
