+++ src/Auth0Provider.jsx
@@ -1,6 +1,16 @@
 import React from "react";
 import { Auth0Provider } from "@auth0/auth0-react";
-import { auth0Config } from "./auth0-config";
+
+export const auth0Config = {
+  domain: import.meta.env.VITE_AUTH0_DOMAIN || "your-domain.auth0.com",
+  clientId: import.meta.env.VITE_AUTH0_CLIENT_ID || "your-client-id",
+  authorizationParams: {
+    redirect_uri: window.location.origin,
+    audience:
+      import.meta.env.VITE_AUTH0_AUDIENCE || "https://your-api-identifier",
+      scope: "openid profile email",
+  },
+};
 
 const Auth0ProviderWrapper = ({ children }) => {
   const onRedirectCallback = (appState) => {
